LayoutTests/ietestcenter/Javascript/15.2.3.3-4-214-expected.txt

15.2.3.3-4-214 (TestCases/chapter15/15.2/15.2.3/15.2.3.3/15.2.3.3-4-214.js) Object.getOwnPropertyDescriptor returns data desc (all false) for properties on built-ins (RegExp.prototype.ignoreCase)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
